requirements.txt
pygame>=2.0.1
python-dotenv>=0.16.0